meilleurs gestion des modificateur dans printf | libft : Ajout des fonciton : strtrim_full, strsub_to, rassemblement des fonciton putnbr, ajout des countnb et abs par type

+++ libft/src/math/ft_abs.c
@@ -3,7 +3,7 @@
 
 unsigned long long ft_abs_ll(long long n)
 {
-	unsigned long	res;
+	unsigned long long	res;
 	
 	if (n < 0)
 	{
